+++ package.json
@@ -11,12 +11,13 @@
     "@gridsome/remark-prismjs": "^0.2.0",
     "@gridsome/source-filesystem": "^0.6.0",
     "@gridsome/transformer-remark": "^0.3.0",
+    "autoprefixer": "^10.4.0",
     "gridsome": "^0.7.0",
-    "gridsome-plugin-netlify-cms": "^1.0.9",
-    "netlify-cms": "^2.10.139"
+    "gridsome-plugin-tailwindcss": "^3.0.1",
+    "gridsome-source-notion": "^1.1.1"
   },
   "devDependencies": {
-    "node-sass": "^4.12.0",
-    "sass-loader": "^8.0.0"
+    "gridsome-plugin-windicss": "^1.0.4",
+    "windicss": "^3.3.0"
   }
 }
