--- src/admin/config.yml
@@ -1,40 +1,0 @@
-backend:
-  name: github
-  repo: michaelhthomas/michaelhthomas.github.io
-  branch: main # Branch to update (optional; defaults to master)
-  site_domain: clever-tereshkova-f61f09.netlify.app
-
-media_folder: "static/uploads"
-public_folder: "/uploads"
-
-publish_mode: editorial_workflow
-
-collections:
-  - name: "posts"
-    label: "Posts"
-    label_singular: "Post"
-    folder: "content/posts"
-    path: "{{year}}-{{month}}-{{day}}-{{slug}}/{{slug}}/{{slug}}"
-    media_folder: ''
-    public_folder: ''
-    create: true
-    slug: "{{slug}}"
-    fields:
-      - label: "Title"
-        name: "title"
-        widget: "string"
-      - label: "Description"
-        name: "description"
-        widget: "string"
-      - label: "Cover Image"
-        name: "cover_image"
-        widget: "image"
-      - label: "Tags"
-        name: "tags"
-        widget: "list"
-      - label: "Publish Date"
-        name: "date"
-        widget: "datetime"
-      - label: "Body"
-        name: "body"
-        widget: "markdown"
