--- src/assets/style/_code.scss
@@ -1,125 +1,0 @@
-/* PrismJS 1.15.0
-https://prismjs.com/download.html#themes=prism&languages=markup+css+clike+javascript */
-
-/**
- * Lucario Theme originally by Raphael Amorim [@raphamorim]
- * https://github.com/raphamorim/lucario
- *
- * Ported for PrismJS by Christopher Kapic [@christopher-kapic]
- */
-
- code[class*="language-"],
- pre[class*="language-"] {
-	 color: #f8f8f2;
-	 background: none;
-	 text-shadow: 0 1px rgba(0, 0, 0, 0.3);
-	 font-family: Monaco, Consolas, 'Andale Mono', 'Ubuntu Mono', monospace;
-	 text-align: left;
-	 white-space: pre;
-	 word-spacing: normal;
-	 word-break: normal;
-	 word-wrap: normal;
-	 line-height: 1.5;
-	 -moz-tab-size: 4;
-	 -o-tab-size: 4;
-	 tab-size: 4;
-	 -webkit-hyphens: none;
-	 -moz-hyphens: none;
-	 -ms-hyphens: none;
-	 hyphens: none;
- }
- 
- /* Code blocks */
- pre[class*="language-"] {
-	 padding: 1em;
-	 margin: .5em 0;
-	 overflow: auto;
-	 border-radius: 0.3em;
- }
- 
- :not(pre) > code[class*="language-"],
- pre[class*="language-"] {
-	 background: #263E52;
- }
- 
- /* Inline code */
- :not(pre) > code[class*="language-"] {
-	 padding: .1em;
-	 border-radius: .3em;
-	 white-space: normal;
- }
- 
- .token.comment,
- .token.prolog,
- .token.doctype,
- .token.cdata {
-	 color: #5c98cd;
- }
- 
- .token.punctuation {
-	 color: #f8f8f2;
- }
- 
- .namespace {
-	 opacity: .7;
- }
- 
- .token.property,
- .token.tag,
- .token.constant,
- .token.symbol,
- .token.deleted {
-	 color: #F05E5D;
- }
- 
- .token.boolean,
- .token.number {
-	 color: #BC94F9;
- }
- 
- .token.selector,
- .token.attr-name,
- .token.string,
- .token.char,
- .token.builtin,
- .token.inserted {
-	 color: #FCFCD6;
- }
- 
- .token.operator,
- .token.entity,
- .token.url,
- .language-css .token.string,
- .style .token.string,
- .token.variable {
-	 color: #f8f8f2;
- }
- 
- .token.atrule,
- .token.attr-value,
- .token.function,
- .token.class-name {
-	 color: #66D8EF;
- }
- 
- .token.keyword {
-	 color: #6EB26E;
- }
- 
- .token.regex,
- .token.important {
-	 color: #F05E5D;
- }
- 
- .token.important,
- .token.bold {
-	 font-weight: bold;
- }
- 
- .token.italic {
-	 font-style: italic;
- }
- 
- .token.entity {
-	 cursor: help;
- }
