--- src/assets/style/_content-box.scss
@@ -1,11 +1,0 @@
-// Used to wrap content inside a nice box.
-.content-box {
-	background-color: var(--bg-content-color);
-	max-width: var(--content-width);
-	margin: 0 auto;
-	transition: background-color .6s;
-	padding: var(--space);
-  border-radius: var(--radius);
-  box-shadow: 1px 1px 5px 0 rgba(0,0,0,.02), 1px 1px 15px 0 rgba(0,0,0,.03);
-  transition: transform .3s, background-color .3s, box-shadow .6s;
-}
