--- src/assets/style/_utils.scss
@@ -1,7 +1,0 @@
-.text-center {
-	text-align: center;
-}
-
-.space-bottom {
-	margin-bottom: var(--space);
-}
