--- src/assets/style/index.scss
@@ -1,7 +1,0 @@
-@import 'reset';
-@import 'variables';
-@import 'typography';
-@import 'base';
-@import 'code';
-@import 'content-box';
-@import 'utils';
