+++ src/components/Logo.vue
@@ -14,18 +14,3 @@ query {
 }
 </static-query>
 
-<style lang="scss">
-.logo {
-	text-decoration: none;
-  color: var(--body-color)!important;
-  font-size: .9em;
-
-	&__image {
-		vertical-align: middle;
-		border-radius: 99px;
-		height: 40px;
-		width: 40px;
-		margin-right: .5em;
-	}
-}
-</style>
