+++ src/components/PostMeta.vue
@@ -1,6 +1,6 @@
 <template>
    <div class="post-meta">
-      Posted {{ post.date_published }}.
+      Posted {{ (post.properties.date_published) ? post.properties.date_published.date.start : "" }}.
       <template v-if="post.timeToRead">
         <strong>{{ post.timeToRead }} min read.</strong>
       </template>
@@ -12,10 +12,3 @@ export default {
   props: ['post']
 }
 </script>
-
-<style lang="scss">
-.post-meta {
-  font-size: .8em;
-  opacity: .8;
-}
-</style>
