+++ src/components/PostTags.vue
@@ -11,20 +11,3 @@ export default {
   props: ['post']
 }
 </script>
-
-<style lang="scss">
-.post-tags {
-  margin: 1em 0 0;
-
-  &__link {
-  	margin-right: .7em;
-  	font-size: .8em;
-  	color: currentColor;
-  	text-decoration: none;
-  	background-color: var(--bg-color);
-  	color: currentColor!important; //Todo: remove important;
-  	padding: .5em;
-  	border-radius: var(--radius);
-  }
-}
-</style>
