+++ src/main.js
@@ -1,6 +1,3 @@
-// Import main css
-import '~/assets/style/index.scss'
-
 // Import default layout so we don't need to import it to every page
 import DefaultLayout from '~/layouts/Default.vue'
