+++ src/pages/Blog.vue
@@ -1,0 +1,50 @@
+<template>
+  <Layout :show-logo="false">
+    <!-- Author intro -->
+    <Author :show-title="true" />
+
+    <!-- List posts -->
+    <div class="posts">
+      <PostCard v-for="edge in $page.posts.edges" :key="edge.node.id" :post="edge.node"/>
+    </div>
+
+  </Layout>
+</template>
+
+<page-query>
+query {
+  posts: allPost(sortBy: "date_published", order: DESC) {
+    edges {
+      node {
+        id
+        title
+        date_published (format: "MMMM D, YYYY")
+        timeToRead
+        description
+        cover_image (width: 770, height: 380, blur: 10)
+        path
+        tags {
+          id
+          title
+          path
+        }
+      }
+    }
+  }
+}
+</page-query>
+
+<script>
+import Author from '~/components/Author.vue'
+import PostCard from '~/components/PostCard.vue'
+
+export default {
+  components: {
+    Author,
+    PostCard
+  },
+  metaInfo: {
+    title: 'Blog'
+  }
+}
+</script>
