+++ src/pages/Index.vue
@@ -3,36 +3,8 @@
     <!-- Author intro -->
     <Author :show-title="true" />
 
-    <!-- List posts -->
-    <div class="posts">
-      <PostCard v-for="edge in $page.posts.edges" :key="edge.node.id" :post="edge.node"/>
-    </div>
-
   </Layout>
 </template>
-
-<page-query>
-query {
-  posts: allPost(sortBy: "date_published", order: DESC) {
-    edges {
-      node {
-        id
-        title
-        date_published (format: "MMMM D, YYYY")
-        timeToRead
-        description
-        cover_image (width: 770, height: 380, blur: 10)
-        path
-        tags {
-          id
-          title
-          path
-        }
-      }
-    }
-  }
-}
-</page-query>
 
 <script>
 import Author from '~/components/Author.vue'
@@ -44,7 +16,7 @@ export default {
     PostCard
   },
   metaInfo: {
-    title: 'Hello, world!'
+    title: 'Home'
   }
 }
 </script>
