+++ windi.config.js
@@ -1,0 +1,12 @@
+module.exports = {
+  purge: {
+    enabled: process.env.NODE_ENV === 'production',
+    content: [
+      './src/**/*.vue',
+      './src/**/*.js',
+      './src/**/*.jsx',
+      './src/**/*.pug',
+      './src/**/*.md'
+    ],
+  },
+}
